+++ medusa/failed_history.py
@@ -1,30 +1,14 @@
 # coding=utf-8
-# Author: Tyler Fenby <tylerfenby@gmail.com>
 
-#
-# This file is part of Medusa.
-#
-# Medusa is free software: you can redistribute it and/or modify
-# it under the terms of the GNU General Public License as published by
-# the Free Software Foundation, either version 3 of the License, or
-# (at your option) any later version.
-#
-# Medusa is distributed in the hope that it will be useful,
-# but WITHOUT ANY WARRANTY; without even the implied warranty of
-# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
-# GNU General Public License for more details.
-#
-# You should have received a copy of the GNU General Public License
-# along with Medusa. If not, see <http://www.gnu.org/licenses/>.
 """failed history code."""
-import re
 
+import re
 from datetime import datetime, timedelta
 
+from medusa.exceptions import EpisodeNotFoundException
 from . import db, logger
 from .common import FAILED, Quality, WANTED, statusStrings
 from .helper.common import episode_num
-from .helper.exceptions import EpisodeNotFoundException
 from .show.history import History
