+++ medusa/failed_processor.py
@@ -1,22 +1,7 @@
 # coding=utf-8
-# Author: Tyler Fenby <tylerfenby@gmail.com>
-#
-# This file is part of Medusa.
-#
-# Medusa is free software: you can redistribute it and/or modify
-# it under the terms of the GNU General Public License as published by
-# the Free Software Foundation, either version 3 of the License, or
-# (at your option) any later version.
-#
-# Medusa is distributed in the hope that it will be useful,
-# but WITHOUT ANY WARRANTY; without even the implied warranty of
-# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
-# GNU General Public License for more details.
-#
-# You should have received a copy of the GNU General Public License
 
+from medusa.exceptions import FailedPostProcessingFailedException
 from . import app, logger, show_name_helpers
-from .helper.exceptions import FailedPostProcessingFailedException
 from .name_parser.parser import InvalidNameException, InvalidShowException, NameParser
 from .search.queue import FailedQueueItem
