+++ medusa/media/generic.py
@@ -1,28 +1,12 @@
 # coding=utf-8
-# This file is part of Medusa.
-#
-
-#
-# Medusa is free software: you can redistribute it and/or modify
-# it under the terms of the GNU General Public License as published by
-# the Free Software Foundation, either version 3 of the License, or
-# (at your option) any later version.
-#
-# Medusa is distributed in the hope that it will be useful,
-# but WITHOUT ANY WARRANTY; without even the implied warranty of
-# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
-# GNU General Public License for more details.
-#
-# You should have received a copy of the GNU General Public License
-# along with Medusa. If not, see <http://www.gnu.org/licenses/>.
 
 from abc import abstractmethod
 from mimetypes import guess_type
 from os.path import isfile, join, normpath
 
+from medusa.exceptions import MultipleShowObjectsException
 from .. import app
 from ..helper.common import try_int
-from ..helper.exceptions import MultipleShowObjectsException
 from ..show.show import Show
