+++ medusa/search/core.py
@@ -1,22 +1,7 @@
 # coding=utf-8
-# Author: Nic Wolfe <nic@wolfeden.ca>
-#
-#
-# This file is part of Medusa.
-#
-# Medusa is free software: you can redistribute it and/or modify
-# it under the terms of the GNU General Public License as published by
-# the Free Software Foundation, either version 3 of the License, or
-# (at your option) any later version.
-#
-# Medusa is distributed in the hope that it will be useful,
-# but WITHOUT ANY WARRANTY; without even the implied warranty of
-# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
-# GNU General Public License for more details.
-#
-# You should have received a copy of the GNU General Public License
-# along with Medusa. If not, see <http://www.gnu.org/licenses/>.
+
 """Search core module."""
+
 import datetime
 import errno
 import os
@@ -52,14 +37,12 @@ from medusa.common import (
     SNATCHED_PROPER,
     UNKNOWN,
 )
+from medusa.exceptions import AuthException
 from medusa.helper.common import (
     enabled_providers,
     episode_num,
 )
-from medusa.helper.exceptions import (
-    AuthException,
-    ex,
-)
+from medusa.helper.exceptions import ex
 from medusa.providers import sorted_provider_list
 from medusa.providers.generic_provider import GenericProvider
