+++ medusa/server/web/manage/handler.py
@@ -9,6 +9,10 @@ import re
 
 from tornroutes import route
 
+from medusa.exceptions import (
+    CantRefreshShowException,
+    CantUpdateShowException,
+)
 from ..core import PageTemplate, WebRoot
 from ..home import Home
 from .... import app, db, helpers, logger, network_timezones, post_processor, sbdatetime, subtitles, ui
@@ -17,10 +21,6 @@ from ....common import (
 )
 from ....helper.common import (
     episode_num, try_int,
-)
-from ....helper.exceptions import (
-    CantRefreshShowException,
-    CantUpdateShowException,
 )
 from ....helpers import is_media_file
 from ....network_timezones import app_timezone
