+++ medusa/show/recommendations/trakt.py
@@ -1,29 +1,17 @@
 # coding=utf-8
-#
-# This file is part of Medusa.
-#
-# Medusa is free software: you can redistribute it and/or modify
-# it under the terms of the GNU General Public License as published by
-# the Free Software Foundation, either version 3 of the License, or
-# (at your option) any later version.
-#
-# Medusa is distributed in the hope that it will be useful,
-# but WITHOUT ANY WARRANTY; without even the implied warranty of
-# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
-# GNU General Public License for more details.
-#
-# You should have received a copy of the GNU General Public License
-# along with Medusa. If not, see <http://www.gnu.org/licenses/>.
+
 from __future__ import unicode_literals
 
 import requests
 from simpleanidb import Anidb
 from traktor import (TokenExpiredException, TraktApi, TraktException)
 from tvdbapiv2 import (ApiClient, AuthenticationApi, SeriesApi)
+
+from medusa.exceptions import MultipleShowObjectsException
+from medusa.helper.exceptions import ex
 from .recommended import RecommendedShow
 from ... import app, logger
 from ...helper.common import try_int
-from ...helper.exceptions import MultipleShowObjectsException, ex
 from ...indexers.indexer_config import INDEXER_TVDBV2
