+++ medusa/show/show.py
@@ -1,27 +1,16 @@
 # coding=utf-8
-# This file is part of Medusa.
-#
-
-#
-# Medusa is free software: you can redistribute it and/or modify
-# it under the terms of the GNU General Public License as published by
-# the Free Software Foundation, either version 3 of the License, or
-# (at your option) any later version.
-#
-# Medusa is distributed in the hope that it will be useful,
-# but WITHOUT ANY WARRANTY; without even the implied warranty of
-# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
-# GNU General Public License for more details.
-#
-# You should have received a copy of the GNU General Public License
-# along with Medusa. If not, see <http://www.gnu.org/licenses/>.
 
 from datetime import date
 
+from medusa.exceptions import (
+    CantRefreshShowException,
+    CantRemoveShowException,
+    MultipleShowObjectsException,
+)
+from medusa.helper.exceptions import ex
 from .. import app
 from ..common import Quality, SKIPPED, WANTED
 from ..db import DBConnection
-from ..helper.exceptions import CantRefreshShowException, CantRemoveShowException, MultipleShowObjectsException, ex
 
 
 class Show(object):
