+++ medusa/tv/cache.py
@@ -1,21 +1,7 @@
 # coding=utf-8
-# Author: Nic Wolfe <nic@wolfeden.ca>
-#
-# This file is part of Medusa.
-#
-# Medusa is free software: you can redistribute it and/or modify
-# it under the terms of the GNU General Public License as published by
-# the Free Software Foundation, either version 3 of the License, or
-# (at your option) any later version.
-#
-# Medusa is distributed in the hope that it will be useful,
-# but WITHOUT ANY WARRANTY; without even the implied warranty of
-# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
-# GNU General Public License for more details.
-#
-# You should have received a copy of the GNU General Public License
-# along with Medusa. If not, see <http://www.gnu.org/licenses/>.
+
 """tv_cache code."""
+
 from __future__ import unicode_literals
 
 import itertools
@@ -28,8 +14,8 @@ from medusa import (
     logger,
     show_name_helpers,
 )
+from medusa.exceptions import AuthException
 from medusa.helper.common import episode_num
-from medusa.helper.exceptions import AuthException
 from medusa.name_parser.parser import (
     InvalidNameException,
     InvalidShowException,
