+++ tests/legacy/show/show_tests.py
@@ -1,18 +1,4 @@
 # coding=utf-8
-# This file is part of Medusa.
-#
-# Medusa is free software: you can redistribute it and/or modify
-# it under the terms of the GNU General Public License as published by
-# the Free Software Foundation, either version 3 of the License, or
-# (at your option) any later version.
-#
-# Medusa is distributed in the hope that it will be useful,
-# but WITHOUT ANY WARRANTY; without even the implied warranty of
-# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
-# GNU General Public License for more details.
-#
-# You should have received a copy of the GNU General Public License
-# along with Medusa. If not, see <http://www.gnu.org/licenses/>.
 
 """Test shows."""
 
@@ -22,9 +8,10 @@ import unittest
 
 from medusa import app
 from medusa.common import Quality
-from medusa.helper.exceptions import MultipleShowObjectsException
+from medusa.exceptions import MultipleShowObjectsException
 from medusa.show.show import Show
 from medusa.tv import Series
+
 from six import iteritems
