+++ tests/legacy/ssl_sni_tests.py
@@ -1,28 +1,16 @@
 # coding=UTF-8
-# Author: Dustyn Gibson <miigotu@gmail.com>
-#
-# This file is part of Medusa.
-#
-# Medusa is free software: you can redistribute it and/or modify
-# it under the terms of the GNU General Public License as published by
-# the Free Software Foundation, either version 3 of the License, or
-# (at your option) any later version.
-#
-# Medusa is distributed in the hope that it will be useful,
-# but WITHOUT ANY WARRANTY; without even the implied warranty of
-# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
-# GNU General Public License for more details.
-#
-# You should have received a copy of the GNU General Public License
-# along with Medusa. If not, see <http://www.gnu.org/licenses/>.
+
 """SSL SNI Tests."""
+
 from __future__ import print_function
 
 import unittest
 
 import certifi
+
 import medusa.providers as providers
 from medusa.helper.exceptions import ex
+
 import requests
