+++ views/editShow.mako
@@ -4,7 +4,6 @@
     from medusa import app, common
     from medusa.common import SKIPPED, WANTED, UNAIRED, ARCHIVED, IGNORED, SNATCHED, SNATCHED_PROPER, SNATCHED_BEST, FAILED
     from medusa.common import statusStrings
-    from medusa.helper import exceptions
     from medusa.indexers.indexer_api import indexerApi
     from medusa.indexers.indexer_config import mappings
     from medusa import scene_exceptions
